Add snapcraft.yml
Adds the snapcraft.yaml needed to build the snap package for ytcc.
Unfortunately, the resulting snap does not work properly and ultimately
I gave up trying to fix the issues. Added the the YAML file despite of
these problems to pick up snap packaging for ytcc when snapcraft is more
mature.