Add raise when adding video to unknown playlist
When adding using database.add_videos() to add to a playlist that
doesn't exist, a TypeError is raised, because of indexing None.
Example below:

Traceback (most recent call last):
  File "import_db.py", line 54, in <module>
    ytcc.database.add_videos(l, ytcc.core.Playlist(current_pl, url))
  File "/ytcc/ytcc/ytcc/database.py",
    playlist_id = cursor.fetchone()["id"]
TypeError: 'NoneType' object is not subscriptable

This commit adds a check to the value of cursor.fetchone(). If it is
None, a PlaylistDoesNotExistException is then raised since there is no
such playlist.