Fix cleanup removing too many videos
The cleanup command removed all watched videos. On the next update they
would be synced again and be available as unwatched videos.
A --keep option was added to specify how many videos are kept on every
playlist. If that option is not specified, the max_update_backlog
setting is used as default.