Remove subshell
fzf supports field index expression, that trims the value by default, so a subshell is not needed.

It also has the advantage not to rely on shell syntax, because since `$SHELL` variable is not set, `fzf` use the user shell to execute the preview, and it can have wrong result. I was impacted as a fish user.